move sqs/queue functions into dedicated module
