+++ cloudigrade/util/aws/__init__.py
@@ -11,6 +11,8 @@ from botocore.exceptions import ClientError
 from django.conf import settings
 from django.utils.translation import gettext as _
 
+from util.aws.sqs import delete_message_from_queue, extract_sqs_message, \
+    receive_message_from_queue
 from util.exceptions import (AwsSnapshotCopyLimitError,
                              AwsSnapshotNotOwnedError,
                              AwsVolumeError,
@@ -518,59 +520,6 @@ def _verify_policy_action(session, action):
         return _handle_dry_run_response_exception(action, e)
 
 
-def receive_message_from_queue(queue_url):
-    """
-    Get message objects from SQS Queue object.
-
-    Args:
-        queue_url (str): The AWS assigned URL for the queue.
-
-    Returns:
-        list[Message]: A list of message objects.
-
-    """
-    region = settings.SQS_DEFAULT_REGION
-    sqs_queue = boto3.resource('sqs', region_name=region).Queue(queue_url)
-
-    messages = sqs_queue.receive_messages(
-        MaxNumberOfMessages=10,
-        WaitTimeSeconds=10
-    )
-
-    return messages
-
-
-def delete_message_from_queue(queue_url, messages):
-    """
-    Delete message objects from SQS queue.
-
-    Args:
-        queue_url (str): The AWS assigned URL for the queue.
-        messages (list[Message]): A list of message objects to delete.
-
-    Returns:
-        dict: The response from the delete call.
-
-    """
-    region = settings.SQS_DEFAULT_REGION
-    sqs_queue = boto3.resource('sqs', region_name=region).Queue(queue_url)
-
-    messages_to_delete = [
-        {
-            'Id': message.message_id,
-            'ReceiptHandle': message._receipt_handle
-        }
-        for message in messages
-    ]
-
-    response = sqs_queue.delete_messages(
-        Entries=messages_to_delete
-    )
-
-    # TODO: Deal with success/failure of message deletes
-    return response
-
-
 def get_object_content_from_s3(bucket, key, compression='gzip'):
     """
     Get the file contents from an S3 object.
@@ -598,25 +547,6 @@ def get_object_content_from_s3(bucket, key, compression='gzip'):
         logger.error(_('Unsupported compression format'))
 
     return content
-
-
-def extract_sqs_message(message, service='s3'):
-    """
-    Parse SQS message for service-specific content.
-
-    Args:
-        message (boto3.SQS.Message): The Message object.
-        service (str): The AWS service the message refers to.
-
-    Returns:
-        list(dict): List of message records.
-
-    """
-    extracted_records = []
-    message_body = json.loads(message.body)
-    for record in message_body.get('Records', []):
-        extracted_records.append(record[service])
-    return extracted_records
 
 
 def rewrap_aws_errors(original_function):
